kash: eliminating warnings (gcc/darwin).
git-svn-id: http://svn.netlabs.org/repos/kbuild/trunk@2312 8c291193-6614-0410-a08d-adc6e07170bb

+++ src/kash/parser.c
@@ -1345,7 +1345,7 @@ badsub:			synerror(psh, "Bad substitution");
 		*(stackblock(psh) + typeloc) = subtype | flags;
 		if (subtype != VSNORMAL) {
 			varnest++;
-			if (varnest >= maxnest) {
+			if (varnest >= (int)maxnest) {
 				dblquotep = ckrealloc(psh, dblquotep, maxnest / 8);
 				dblquotep[(maxnest / 32) - 1] = 0;
 				maxnest += 32;
