+++ src/kash/setmode.c
@@ -367,7 +367,7 @@ addcmd(set, op, who, oparg, mask)
 	case '-':
 	case 'X':
 		set->cmd = op;
-		set->bits = (who ? who : mask) & oparg;
+		set->bits = (who ? (unsigned int)who : mask) & (unsigned int)oparg;
 		break;
 
 	case 'u':
